Added option to preserve GraphViz file

+++ diagrams/__init__.py
@@ -83,6 +83,7 @@ class Diagram:
         graph_attr: dict = {},
         node_attr: dict = {},
         edge_attr: dict = {},
+        preserve_graphviz_file=false
     ):
         """Diagram represents a global diagrams context.
 
@@ -138,7 +139,8 @@ class Diagram:
     def __exit__(self, exc_type, exc_value, traceback):
         self.render()
         # Remove the graphviz file leaving only the image.
-        os.remove(self.filename)
+        if (not self.preserve_graphviz_file):
+            os.remove(self.filename)
         setdiagram(None)
 
     def _repr_png_(self):
@@ -487,7 +489,8 @@ class Edge:
                 self._attrs = o.attrs.copy()
                 result.append(o)
             else:
-                result.append(Edge(o, forward=forward, reverse=reverse, **self._attrs))
+                result.append(
+                    Edge(o, forward=forward, reverse=reverse, **self._attrs))
         return result
 
     def connect(self, other: Union["Node", "Edge", List["Node"]]):
